backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
websockets
requests
sounddevice
numpy
